+++ lib/td_dq/implementations/implementation.ex
@@ -50,6 +50,12 @@ defmodule TdDq.Implementations.Implementation do
 
     has_many(:data_structures, ImplementationStructure, where: [deleted_at: nil])
 
+    has_many(:dataset_structures, ImplementationStructure,
+      where: [deleted_at: nil, type: :dataset]
+    )
+
+    has_many(:dataset_sources, through: [:dataset_structures, :source])
+
     timestamps(type: :utc_datetime)
   end